Fixed compiling warning cause I suck

+++ source/planner/locModule.h
@@ -138,7 +138,7 @@ private:
 	double getGPSLongitude(void);
 	double getGPSLatitude(void);
 	double getHeading(void);
-	bool getEncoderValue(long int *pValue);
+	bool getEncoderValue(int *pValue);
 	double getSteeringAngleRad(void);
   int getMotorSetPoint(void);
   int getAccelerometerReadings (float *ddx, float *ddy, float *ddz);
